fix: assert the error.data response
