+++ v-next/hardhat/src/internal/builtin-plugins/network-manager/edr/type-validation.ts
@@ -1,0 +1,25 @@
+import type { DebugTraceResult } from "@ignored/edr-optimism";
+
+import { isObject } from "@ignored/hardhat-vnext-utils/lang";
+
+export function isDebugTraceResult(
+  result: unknown,
+): result is DebugTraceResult {
+  return (
+    isObject(result) &&
+    "pass" in result &&
+    "gasUsed" in result &&
+    "structLogs" in result
+  );
+}
+
+interface EdrProviderErrorData {
+  data: string;
+  transactionHash?: string;
+}
+
+export function isEdrProviderErrorData(
+  errorData: unknown,
+): errorData is EdrProviderErrorData {
+  return isObject(errorData) && "data" in errorData;
+}
